Refactor SearchFragment
