+++ app/src/main/res/layout/activity_search.xml
@@ -2,10 +2,10 @@
 <fragment xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/search_fragment"
-    android:name="org.jellyfin.androidtv.search.MySearchFragment"
+    android:name="org.jellyfin.androidtv.search.SearchFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ByLetterActivity"
+    tools:context=".search.SearchActivity"
     tools:deviceIds="tv"
     tools:ignore="MergeRootFrame"
     tools:layout="@layout/activity_search" />
